requirements.txt
slack-bolt==1.18.1
slack_sdk==3.27.1
python-dotenv~=1.0.1
langchain~=0.1.12
langchain-openai~=0.1.4
requests~=2.31.0
pydantic~=2.6.4
asyncpg~=0.29.0
langchainhub~=0.1.15
psycopg~=3.1.18
psycopg2~=2.9.9
pgvector~=0.2.5
fastapi~=0.109.2
pydantic~=2.6.1
python-dotenv~=1.0.1
databases~=0.8.0
uvicorn~=0.29.0
asyncpg~=0.29.0
